# docs/networkdisk-apifox-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NetworkDisk 本地接口",
    "version": "1.0.0",
    "description": "当前后端实际开放的 9 个 HTTP 接口。浏览器 POST 前先请求 /api/v1/auth/csrf，并在 X-XSRF-TOKEN 请求头回传 CSRF token。推荐顺序：发送邮箱验证码 → 注册（已有账号可跳过）→ 获取图片验证码 → 登录 → GET /api/v1/auth/me 确认会话 → 修改密码或退出登录。图片验证码与用户名及 X-Device-Id 绑定，登录后立即失效。登录成功后 JWT 通过 HttpOnly Cookie 下发，不会出现在 JSON 响应中；受保护请求自动携带 Cookie。修改密码后旧 Token 失效；退出登录撤销当前 Token 并清除认证及 CSRF Cookie。退出后再次登录如遇无业务错误码的 403，应重新获取 CSRF token 后重试。"
  },
  "servers": [
    {
      "url": "http://localhost:9090",
      "description": "本地 Spring Boot 服务"
    }
  ],
  "tags": [
    { "name": "系统", "description": "服务状态" },
    { "name": "注册", "description": "邮箱验证码和账号注册" },
    { "name": "登录", "description": "图片验证码、登录和令牌管理" },
    { "name": "账户安全", "description": "修改密码与退出登录" }
  ],
  "paths": {
    "/": {
      "get": {
        "tags": ["系统"],
        "summary": "检查服务状态",
        "operationId": "getSystemStatus",
        "responses": {
          "200": {
            "description": "后端已启动",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SystemResult" },
                "example": {
                  "success": true,
                  "code": "OK",
                  "message": "操作成功",
                  "data": { "service": "network-disk-backend", "status": "running" }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/email-code": {
      "post": {
        "tags": ["注册"],
        "summary": "发送注册邮箱验证码",
        "description": "向尚未注册的邮箱发送 6 位数字验证码；验证码有效期为 60 秒。使用注册接口提交收到的验证码。要完成真实注册，请保持邮件发送功能开启。",
        "operationId": "sendRegistrationEmailCode",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EmailCodeRequest" },
              "example": { "email": "new-user@example.com" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "发送成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StringResult" },
                "example": {
                  "success": true,
                  "code": "OK",
                  "message": "操作成功",
                  "data": "验证码已发送"
                }
              }
            }
          },
          "400": {
            "description": "邮箱参数错误或邮箱已注册（INVALID_PARAM、EMAIL_EXISTS）",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResult" } } }
          },
          "429": {
            "description": "发送频率限制（IP_RATE_LIMITED、EMAIL_COOLDOWN、EMAIL_DAILY_LIMITED、SEND_IN_PROGRESS）",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResult" } } }
          },
          "503": {
            "description": "Redis 或邮件服务不可用（REDIS_UNAVAILABLE、MAIL_SEND_FAILED）",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResult" } } }
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": ["注册"],
        "summary": "注册账号",
        "description": "必须先通过 /api/v1/auth/email-code 获取邮箱验证码。邮箱入库时转为小写；用户名仅去除首尾空格并保留原始大小写；密码以 BCrypt 哈希保存。confirmPassword 只在前端校验，不提交给后端。注册成功不自动登录。",
        "operationId": "registerUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" },
              "example": {
                "email": "new-user@example.com",
                "nickName": "Admin",
                "password": "ExamplePassword123",
                "emailCode": "123456"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "注册成功，data 为用户 ID 字符串",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StringResult" },
                "example": { "success": true, "code": "OK", "message": "操作成功", "data": "1" }
              }
            }
          },
          "400": {
            "description": "参数、验证码或唯一性校验失败（INVALID_PARAM、CODE_EXPIRED、INVALID_EMAIL_CODE、EMAIL_EXISTS、NICKNAME_EXISTS、REGISTER_CONFLICT）",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResult" } } }
          },
          "429": {
            "description": "同一邮箱的注册请求正在处理（REGISTER_IN_PROGRESS）",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResult" } } }
          },
          "503": {
            "description": "Redis 不可用（REDIS_UNAVAILABLE）",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResult" } } }
          }
        }
      }
    },
    "/api/v1/auth/captcha": {
      "get": {
        "tags": ["登录"],
        "summary": "获取登录图片验证码",
        "description": "返回 data:image/png;base64,... 格式的图片。将完整的 data.image 值复制到浏览器地址栏可查看图片。验证码有效期默认 300 秒，登录验证通过后只可使用一次。获取图片和登录必须使用同一用户名与同一 X-Device-Id；用户名在服务端统一 trim 并转小写。刷新图片会覆盖之前的验证码。",
        "operationId": "getLoginCaptcha",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": true,
            "description": "已注册用户名，3–32 个字符；大小写不影响登录校验",
            "schema": { "type": "string", "minLength": 3, "maxLength": 32 },
            "example": "Admin"
          },
          { "$ref": "#/components/parameters/DeviceId" }
        ],
        "responses": {
          "200": {
            "description": "生成验证码成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CaptchaResult" },
                "example": {
                  "success": true,
                  "code": "OK",
                  "message": "操作成功",
                  "data": {
                    "image": "data:image/png;base64,iVBORw0KGgo...",
                    "prompt": "请输入图片中的4位数字"
                  }
                }
              }
            }
          },
          "400": {
            "description": "用户名或设备标识无效（INVALID_PARAM、INVALID_DEVICE_ID）；缺少请求头时可能由 Spring 直接返回参数错误",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResult" } } }
          },
          "503": {
            "description": "Redis 不可用（REDIS_UNAVAILABLE）",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResult" } } }
          }
        }
      }
    },
    "/api/v1/auth/csrf": {
      "get": {
        "tags": ["登录"],
        "summary": "初始化 CSRF 校验",
        "description": "返回 CSRF token 并设置 XSRF-TOKEN Cookie；所有 POST 请求均需在 X-XSRF-TOKEN 请求头中回传该值。",
        "operationId": "getCsrfToken",
        "responses": {
          "200": {
            "description": "CSRF token 已签发",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/StringResult" } } }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["登录"],
        "summary": "用户名密码登录",
        "description": "每次登录都必须先获取新的图片验证码。仅支持用户名登录，用户名在服务端统一 trim 并转小写用于 Redis 与 MySQL 查询；响应中的 username 为数据库保存的原始大小写。X-Device-Id 必须与获取图片时相同。登录成功后 HS256 JWT 默认有效期 7200 秒，通过 NETWORKDISK_AUTH HttpOnly Cookie 下发；JSON 只返回 userId 和 username。Redis 使用 JWT 的 SHA-256 摘要登记当前有效 Token。浏览器需先获取 CSRF token 并回传 X-XSRF-TOKEN。",
        "operationId": "loginUser",
        "parameters": [
          { "$ref": "#/components/parameters/DeviceId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" },
              "example": {
                "username": "Admin",
                "password": "ExamplePassword123",
                "captchaCode": "2345"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResult" },
                "example": {
                  "success": true,
                  "code": "OK",
                  "message": "操作成功",
                  "data": {
                    "userId": 1,
                    "username": "Admin"
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数或验证码错误（INVALID_PARAM、INVALID_DEVICE_ID、CAPTCHA_INVALID）",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResult" } } }
          },
          "401": {
            "description": "用户名或密码错误（LOGIN_FAILED）",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResult" } } }
          },
          "423": {
            "description": "账号被锁定（ACCOUNT_LOCKED）",
            "headers": { "Retry-After": { "$ref": "#/components/headers/RetryAfter" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResult" } } }
          },
          "429": {
            "description": "IP 或设备登录请求过于频繁（IP_RATE_LIMITED、DEVICE_RATE_LIMITED）",
            "headers": { "Retry-After": { "$ref": "#/components/headers/RetryAfter" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResult" } } }
          },
          "503": {
            "description": "Redis 不可用（REDIS_UNAVAILABLE）",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResult" } } }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": ["登录"],
        "summary": "获取当前登录用户",
        "description": "使用浏览器自动携带的 NETWORKDISK_AUTH Cookie 校验会话，并返回 userId、username。前端路由守卫用于确认登录状态；GET 请求只读，不需要 CSRF token。",
        "operationId": "getCurrentUser",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "当前用户信息",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResult" },
                "example": {
                  "success": true,
                  "code": "OK",
                  "message": "操作成功",
                  "data": { "userId": 1, "username": "Admin" }
                }
              }
            }
          },
          "401": {
            "description": "Cookie 缺失、过期、已撤销或密码已变化（UNAUTHORIZED）",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResult" } } }
          },
          "503": {
            "description": "认证状态服务不可用（AUTH_SERVICE_UNAVAILABLE）",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResult" } } }
          }
        }
      }
    },
    "/api/v1/auth/change-password": {
      "post": {
        "tags": ["账户安全"],
        "summary": "修改当前用户密码",
        "description": "浏览器自动携带 HttpOnly 登录 Cookie。旧密码校验正确后保存新 BCrypt 哈希；成功后清除 Cookie，先前签发的所有 Token 均失效，需要重新登录。",
        "operationId": "changePassword",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChangePasswordRequest" },
              "example": { "currentPassword": "OldPassword123", "newPassword": "NewPassword123" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "密码已修改，原 Token 已失效",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StringResult" },
                "example": { "success": true, "code": "OK", "message": "操作成功", "data": "密码已修改，请重新登录" }
              }
            }
          },
          "400": {
            "description": "参数错误或当前密码错误（INVALID_PARAM、INVALID_CURRENT_PASSWORD）",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResult" } } }
          },
          "401": {
            "description": "Token 缺失、过期、已撤销或密码已变化（UNAUTHORIZED）",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResult" } } }
          },
          "503": {
            "description": "认证状态服务不可用（AUTH_SERVICE_UNAVAILABLE）",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResult" } } }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": ["账户安全"],
        "summary": "退出当前登录",
        "description": "使用当前有效 JWT 和 CSRF token 调用。成功后 Redis 撤销本次请求携带的 Token，并将 NETWORKDISK_AUTH 与 XSRF-TOKEN Cookie 设为过期；同一账号其他设备的 Token 不受此操作影响。重新登录前需要获取新的 CSRF token。",
        "operationId": "logout",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "当前 Token 已撤销",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StringResult" },
                "example": { "success": true, "code": "OK", "message": "操作成功", "data": "已退出登录" }
              }
            }
          },
          "401": {
            "description": "Token 缺失、过期或已撤销（UNAUTHORIZED）",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResult" } } }
          },
          "503": {
            "description": "认证状态服务不可用（AUTH_SERVICE_UNAVAILABLE）",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResult" } } }
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "DeviceId": {
        "name": "X-Device-Id",
        "in": "header",
        "required": true,
        "description": "固定设备 UUID；获取验证码和登录时使用同一个值。可在本地先生成一次 UUID，再保存为 Apifox 本地环境变量。",
        "schema": { "type": "string", "format": "uuid" },
        "example": "123e4567-e89b-42d3-a456-426614174000"
      }
    },
    "headers": {
      "RetryAfter": {
        "description": "建议等待的秒数",
        "schema": { "type": "integer", "minimum": 1 }
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "NETWORKDISK_AUTH",
        "description": "登录接口通过 HttpOnly Cookie 自动设置；浏览器脚本无法读取。"
      }
    },
    "schemas": {
      "EmailCodeRequest": {
        "type": "object",
        "required": ["email"],
        "properties": {
          "email": { "type": "string", "format": "email", "description": "接收验证码的未注册邮箱" }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["email", "nickName", "password", "emailCode"],
        "properties": {
          "email": { "type": "string", "format": "email", "maxLength": 128 },
          "nickName": { "type": "string", "minLength": 3, "maxLength": 32, "description": "注册用户名，服务端保留原始大小写" },
          "password": { "type": "string", "format": "password", "minLength": 6, "maxLength": 64 },
          "emailCode": { "type": "string", "pattern": "^[0-9]{6}$", "description": "邮件中收到的 6 位验证码；业务层强制校验" },
          "checkCode": { "type": "string", "deprecated": true, "description": "预留字段，目前未参与注册校验，不需要传" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password", "captchaCode"],
        "properties": {
          "username": { "type": "string", "minLength": 3, "maxLength": 32 },
          "password": { "type": "string", "format": "password", "minLength": 6, "maxLength": 64 },
          "captchaCode": { "type": "string", "pattern": "^[0-9]{4}$", "description": "本次图片中的 4 位数字" }
        }
      },
      "ChangePasswordRequest": {
        "type": "object",
        "required": ["currentPassword", "newPassword"],
        "properties": {
          "currentPassword": { "type": "string", "format": "password", "description": "当前密码" },
          "newPassword": { "type": "string", "format": "password", "minLength": 6, "maxLength": 64, "description": "新密码" }
        }
      },
      "SystemResult": {
        "type": "object",
        "required": ["success", "code", "message", "data"],
        "properties": {
          "success": { "type": "boolean", "enum": [true] },
          "code": { "type": "string", "example": "OK" },
          "message": { "type": "string", "example": "操作成功" },
          "data": {
            "type": "object",
            "required": ["service", "status"],
            "properties": {
              "service": { "type": "string", "example": "network-disk-backend" },
              "status": { "type": "string", "example": "running" }
            }
          }
        }
      },
      "StringResult": {
        "type": "object",
        "required": ["success", "code", "message", "data"],
        "properties": {
          "success": { "type": "boolean", "enum": [true] },
          "code": { "type": "string", "example": "OK" },
          "message": { "type": "string", "example": "操作成功" },
          "data": { "type": "string" }
        }
      },
      "CaptchaResult": {
        "type": "object",
        "required": ["success", "code", "message", "data"],
        "properties": {
          "success": { "type": "boolean", "enum": [true] },
          "code": { "type": "string", "example": "OK" },
          "message": { "type": "string", "example": "操作成功" },
          "data": {
            "type": "object",
            "required": ["image", "prompt"],
            "properties": {
              "image": { "type": "string", "description": "PNG 图片的 data URL" },
              "prompt": { "type": "string", "example": "请输入图片中的4位数字" }
            }
          }
        }
      },
      "LoginResult": {
        "type": "object",
        "required": ["success", "code", "message", "data"],
        "properties": {
          "success": { "type": "boolean", "enum": [true] },
          "code": { "type": "string", "example": "OK" },
          "message": { "type": "string", "example": "操作成功" },
          "data": {
            "type": "object",
            "required": ["userId", "username"],
            "properties": {
              "userId": { "type": "integer", "format": "int64", "example": 1 },
              "username": { "type": "string", "example": "Admin", "description": "数据库中保存的原始大小写" }
            }
          }
        }
      },
      "ErrorResult": {
        "type": "object",
        "required": ["success", "code", "message", "data"],
        "properties": {
          "success": { "type": "boolean", "enum": [false] },
          "code": { "type": "string", "example": "INVALID_PARAM" },
          "message": { "type": "string", "example": "请检查请求参数" },
          "data": { "nullable": true, "description": "错误时为 null" }
        }
      }
    }
  }
}